tidied code to remove circular dependency
