+++ app/models.py
@@ -4,13 +4,7 @@ from hashlib import md5
 from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
 
-from app import db
-from app import login
-
-
-@login.user_loader
-def load_user(id):
-    return User.query.get(int(id))
+from app.db import db
 
 
 class User(UserMixin, db.Model):
