--- app/support/__init__.py
@@ -1,7 +1,0 @@
-from flask import Blueprint
-
-
-bp = Blueprint('support', __name__)
-
-from app.support import routes, legal_routes
-
