+++ app/templates/index.html
@@ -13,7 +13,7 @@
 
         <div class="row">
             <div class="col-md-8 mx-auto">
-                <p id="no_reg"> Nothing to show here.</p>
+                <p id="no_reg"> Hello World! Thanks for logging in.</p>
             </div>
         </div>
