+++ requirements.txt
@@ -26,6 +26,7 @@ Flask-Testing==0.8.0
 Flask-WTF==0.14.3
 gitdb==4.0.5
 GitPython==3.1.8
+gunicorn==20.0.4
 idna==2.10
 importlib-metadata==2.0.0
 iniconfig==1.0.1
@@ -46,6 +47,8 @@ pycodestyle==2.6.0
 pycparser==2.20
 pyflakes==2.2.0
 pylint==2.6.0
+pylint-flask==0.6
+pylint-plugin-utils==0.6
 PyMySQL==0.10.1
 pyparsing==2.4.7
 pytest==6.0.2
