+++ setup.cfg
@@ -1,6 +1,6 @@
 [pylint]
 persistent=yes
-load-plugins=
+load-plugins=pylint_flask
 jobs=1
 unsafe-load-any-extension=no
 extension-pkg-whitelist=
